fix(chat): promote streaming usage.cost into ResponseMeta.cost_usd_cents
Python port of the TS SDK fix (llm4agents-sdk@2.6.1). The proxy's
streaming endpoint doesn't carry x-cost-usd-cents in the response
headers — by the time headers flush, the round's cost isn't known yet.
The final cost arrives on the terminating SSE chunk's `usage.cost`
field (USD float). The SDK previously captured prompt/completion/
reasoning tokens from that chunk but dropped `cost`, so
`ResponseMeta.cost_usd_cents` stayed None for every streaming round.

Live impact in agent-builder: every meta event reaching the operator
dashboard reported `cost_usd_micros: 0` despite proxy debits.

Fix:
- Capture `chunk.usage.cost` during the SSE loop (already stored in
  chunk_usage; just promote it).
- Before yielding the meta event, replace `round_meta` with a copy
  whose `cost_usd_cents` is set to `cost * 100` (USD → cents) — but
  ONLY when the header-based path didn't already populate it, so
  non-streaming rounds and proxies that DO set the header keep
  working unchanged. Uses `dataclasses.replace` because `ResponseMeta`
  is frozen.
- Widen `ResponseMeta.cost_usd_cents` from `int | None` to
  `float | None`. Streaming completions can produce fractional cents
  (e.g. 0.0035 for a haiku round); integer values from the
  `x-cost-usd-cents` header still satisfy the new type without
  any caller-side changes.

Tests: +3 cases mirroring the TS suite — streaming with usage.cost
only (typical proxy behavior), header wins over usage.cost when both
present, None when neither provides cost. 146/147 Python suite green.

Version bump 2.6.0 → 2.6.1 (semver patch — bug fix, behavior fix only).

+++ llm4agents/chat/types.py
@@ -15,7 +15,13 @@ class ChatMessage(TypedDict, total=False):
 class ResponseMeta:
     request_id: str | None
     model_used: str | None
-    cost_usd_cents: int | None
+    # `float` rather than `int` because streaming rounds promote the
+    # terminating SSE chunk's `usage.cost` (USD) into this field as cents
+    # (× 100), and micro-spend rounds can be fractional cents (e.g. 0.0035
+    # for a haiku completion). The non-streaming path still receives an
+    # integer here via the `x-cost-usd-cents` header — both shapes satisfy
+    # `float | None` without truncation.
+    cost_usd_cents: float | None
     balance_remaining_cents: int | None
     tokens_input: int | None
     tokens_output: int | None
